Renamed Toolbar to BottomAppBar
Toolbar is too general. It only works on the bottom because fixed the layout gravity to bottom. Also there's no iOS equivalent so clearer to give it the same name as the Android view?!
