+++ NavigationReactNative/src/android/src/main/java/com/navigation/reactnative/BottomAppBarManager.java
@@ -7,21 +7,21 @@ import com.facebook.react.uimanager.ThemedReactContext;
 import com.facebook.react.uimanager.ViewGroupManager;
 import com.facebook.react.uimanager.annotations.ReactProp;
 
-public class ToolbarBottomManager extends ViewGroupManager<ToolbarBottomView> {
+public class BottomAppBarManager extends ViewGroupManager<BottomAppBarView> {
     @NonNull
     @Override
     public String getName() {
-        return "NVToolbarBottom";
+        return "NVBottomAppBar";
     }
 
     @NonNull
     @Override
-    protected ToolbarBottomView createViewInstance(@NonNull ThemedReactContext reactContext) {
-        return new ToolbarBottomView(reactContext);
+    protected BottomAppBarView createViewInstance(@NonNull ThemedReactContext reactContext) {
+        return new BottomAppBarView(reactContext);
     }
 
     @ReactProp(name = "height")
-    public void setHeight(ToolbarBottomView view, double height) {
+    public void setHeight(BottomAppBarView view, double height) {
         view.getLayoutParams().height = (int) PixelUtil.toPixelFromDIP(height);
     }
 }
