+++ NavigationReactNative/src/android/src/main/java/com/navigation/reactnative/BottomAppBarView.java
@@ -9,8 +9,8 @@ import androidx.coordinatorlayout.widget.CoordinatorLayout;
 
 import com.google.android.material.bottomappbar.BottomAppBar;
 
-public class ToolbarBottomView extends BottomAppBar {
-    public ToolbarBottomView(@NonNull Context context) {
+public class BottomAppBarView extends BottomAppBar {
+    public BottomAppBarView(@NonNull Context context) {
         super(context);
         CoordinatorLayout.LayoutParams params = new CoordinatorLayout.LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.WRAP_CONTENT);
         params.gravity = Gravity.BOTTOM;
